requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
dj-database-url
python-dotenv
